Foundation Lib: addressing all remaining compiler warnings.

+++ Foundation/library/Mersenne-1.1/MersenneTwister.h
@@ -215,7 +215,7 @@ inline void MTRand::seed( uint32 *const bigSeed, const uint32 seedLength )
 	initialize(19650218UL);
 	int i = 1;
 	uint32 j = 0;
-	int k = ( N > seedLength ? N : seedLength );
+	int k = (int)( N > seedLength ? N : seedLength );
 	for( ; k; --k )
 	{
 		state[i] =
@@ -376,7 +376,7 @@ inline void MTRand::load( uint32 *const loadArray )
 	uint32 *la = loadArray;
 	int i = N;
 	for( ; i--; *s++ = *la++ ) {}
-	left = *la;
+	left = (int)*la;
 	pNext = &state[N-left];
 }
